Refactored ParamSpecsObjectFactory to param-specs.model
