+++ core/templates/domain/exploration/ExplorationMetadataObjectFactory.ts
@@ -19,11 +19,7 @@
 import {Injectable} from '@angular/core';
 import {ParamChange, ParamChangeBackendDict} from './param-change.model';
 import {ParamChanges} from './param-changes.model';
-import {
-  ParamSpecs,
-  ParamSpecsBackendDict,
-  ParamSpecsObjectFactory,
-} from './ParamSpecsObjectFactory';
+import {ParamSpecs, ParamSpecsBackendDict} from './param-specs.model';
 
 export interface ExplorationMetadataBackendDict {
   title: string;
@@ -113,22 +109,17 @@ export class ExplorationMetadata {
   }
 }
 
-// TODO(#15599): Refactor ExplorationMetadataObjectFactory to a model.ts once
-// ParamSpecsObjectFactory and ParamChangesObjectFactory are refactored
-// into model.ts files.
 @Injectable({
   providedIn: 'root',
 })
 export class ExplorationMetadataObjectFactory {
-  constructor(private paramSpecsObjectFactory: ParamSpecsObjectFactory) {}
-
   createFromBackendDict(
     explorationMetadataBackendDict: ExplorationMetadataBackendDict
   ): ExplorationMetadata {
     const paramChanges = ParamChanges.createFromBackendList(
       explorationMetadataBackendDict.param_changes
     );
-    const paramSpecs = this.paramSpecsObjectFactory.createFromBackendDict(
+    const paramSpecs = ParamSpecs.createFromBackendDict(
       explorationMetadataBackendDict.param_specs
     );
