+++ core/templates/domain/exploration/read-only-exploration-backend-api.service.ts
@@ -22,7 +22,7 @@ import {HttpClient} from '@angular/common/http';
 
 import {AppConstants} from 'app.constants';
 import {ParamChangeBackendDict} from 'domain/exploration/param-change.model';
-import {ParamSpecsBackendDict} from 'domain/exploration/ParamSpecsObjectFactory';
+import {ParamSpecsBackendDict} from 'domain/exploration/param-specs.model';
 import {StateObjectsBackendDict} from 'domain/exploration/StatesObjectFactory';
 import {UrlInterpolationService} from 'domain/utilities/url-interpolation.service';
 import {ExplorationMetadataBackendDict} from './ExplorationMetadataObjectFactory';
