+++ core/templates/pages/exploration-editor-page/exploration-editor-page.component.spec.ts
@@ -29,7 +29,6 @@ import {
 import {BrowserDynamicTestingModule} from '@angular/platform-browser-dynamic/testing';
 import {NgbModal, NgbModalRef} from '@ng-bootstrap/ng-bootstrap';
 import {StateEditorService} from 'components/state-editor/state-editor-properties-services/state-editor.service';
-import {ParamSpecsObjectFactory} from 'domain/exploration/ParamSpecsObjectFactory';
 import {UrlInterpolationService} from 'domain/utilities/url-interpolation.service';
 import {StateEditorRefreshService} from 'pages/exploration-editor-page/services/state-editor-refresh.service';
 import {UserExplorationPermissionsService} from 'pages/exploration-editor-page/services/user-exploration-permissions.service';
@@ -290,7 +289,6 @@ describe('Exploration editor page component', () => {
         InternetConnectivityService,
         PageTitleService,
         LoaderService,
-        ParamSpecsObjectFactory,
         RouterService,
         SiteAnalyticsService,
         StateEditorRefreshService,
