+++ core/templates/pages/exploration-editor-page/modal-templates/metadata-version-history-modal.component.spec.ts
@@ -35,7 +35,7 @@ import {
 } from '../services/version-history.service';
 import {MetadataVersionHistoryModalComponent} from './metadata-version-history-modal.component';
 import {ExplorationMetadata} from 'domain/exploration/ExplorationMetadataObjectFactory';
-import {ParamSpecs} from 'domain/exploration/ParamSpecsObjectFactory';
+import {ParamSpecs} from 'domain/exploration/param-specs.model';
 
 describe('Metadata version history modal', () => {
   let component: MetadataVersionHistoryModalComponent;
