+++ core/templates/pages/exploration-editor-page/services/exploration-param-specs.service.ts
@@ -22,7 +22,7 @@ import {ExplorationPropertyService} from 'pages/exploration-editor-page/services
 import {AlertsService} from 'services/alerts.service';
 import {ChangeListService} from './change-list.service';
 import {LoggerService} from 'services/contextual/logger.service';
-import {ParamSpecs} from 'domain/exploration/ParamSpecsObjectFactory';
+import {ParamSpecs} from 'domain/exploration/param-specs.model';
 
 @Injectable({
   providedIn: 'root',
