+++ core/templates/pages/exploration-editor-page/services/exploration-states.service.ts
@@ -53,7 +53,7 @@ import {Outcome} from 'domain/exploration/outcome.model';
 import {Hint} from 'domain/exploration/hint-object.model';
 import {Solution} from 'domain/exploration/solution.model';
 import {InteractionCustomizationArgs} from 'interactions/customization-args-defs';
-import {ParamSpecs} from 'domain/exploration/ParamSpecsObjectFactory';
+import {ParamSpecs} from 'domain/exploration/param-specs.model';
 import {ParamChange} from 'domain/exploration/param-change.model';
 import {
   SubtitledHtml,
