+++ core/templates/pages/exploration-editor-page/services/version-history.service.spec.ts
@@ -18,7 +18,7 @@
 
 import {TestBed} from '@angular/core/testing';
 import {ExplorationMetadata} from 'domain/exploration/ExplorationMetadataObjectFactory';
-import {ParamSpecs} from 'domain/exploration/ParamSpecsObjectFactory';
+import {ParamSpecs} from 'domain/exploration/param-specs.model';
 import {StateObjectFactory} from 'domain/state/StateObjectFactory';
 import {VersionHistoryService} from './version-history.service';
