+++ core/tests/root-files-config.json
@@ -26,7 +26,6 @@
     "core/templates/components/common-layout-directives/navigation-bars/side-navigation-bar.component.css",
     "core/templates/components/forms/validators/README.md",
     "core/templates/css/README.md",
-    "core/templates/domain/exploration/ParamSpecsObjectFactorySpec.ts",
     "core/templates/domain/exploration/ParamTypeObjectFactorySpec.ts",
     "core/templates/domain/statistics/learner-answer-details.model.ts",
     "core/templates/domain/exploration/TranslatedContentObjectFactorySpec.ts",
